TFP-1201 fjerne kode og konfig for infotrygd ws
