--- domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/ytelse/infotrygd/beregningsgrunnlag/YtelseBeregningsgrunnlag.java
@@ -1,88 +1,0 @@
-package no.nav.foreldrepenger.abakus.registerdata.ytelse.infotrygd.beregningsgrunnlag;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-import no.nav.foreldrepenger.abakus.domene.iay.YtelseGrunnlagBuilder;
-import no.nav.foreldrepenger.abakus.domene.iay.kodeverk.Arbeidskategori;
-import no.nav.foreldrepenger.abakus.kodeverk.YtelseType;
-import no.nav.foreldrepenger.abakus.vedtak.domene.TemaUnderkategori;
-import no.nav.tjeneste.virksomhet.infotrygdberegningsgrunnlag.v1.informasjon.Grunnlag;
-import no.nav.tjeneste.virksomhet.infotrygdberegningsgrunnlag.v1.informasjon.Vedtak;
-import no.nav.vedtak.felles.integrasjon.felles.ws.DateUtil;
-
-public abstract class YtelseBeregningsgrunnlag {
-
-    private final YtelseType type;
-    private final TemaUnderkategori temaUnderkategori;
-    private final LocalDate identdato;
-    private LocalDate fom;
-    private LocalDate tom;
-    private List<YtelseBeregningsgrunnlagVedtak> vedtak;
-
-    YtelseBeregningsgrunnlag(YtelseType type, TemaUnderkategori tuk, Grunnlag grunnlag) {
-        this.type = type;
-        this.temaUnderkategori = tuk;
-        identdato = DateUtil.convertToLocalDate(grunnlag.getIdentdato());
-        if (grunnlag.getPeriode() != null) {
-            fom = DateUtil.convertToLocalDate(grunnlag.getPeriode().getFom());
-            tom = DateUtil.convertToLocalDate(grunnlag.getPeriode().getTom());
-        }
-
-        //Trenger trolig ikke behandlingstema her pga behandlingstema fra infotrygdSak er den vi bruker og de er ulike..
-        lagVedtak(grunnlag.getVedtakListe());
-    }
-
-    private void lagVedtak(List<Vedtak> vedtakListe) {
-        ArrayList<YtelseBeregningsgrunnlagVedtak> vedtaks = new ArrayList<>();
-        for (Vedtak vt : vedtakListe) {
-            vedtaks.add(new YtelseBeregningsgrunnlagVedtak(vt));
-        }
-        vedtak = Collections.unmodifiableList(vedtaks);
-    }
-
-    public YtelseType getType() {
-        return type;
-    }
-
-    public TemaUnderkategori getTemaUnderkategori() {
-        return temaUnderkategori;
-    }
-
-    public LocalDate getIdentdato() {
-        return identdato;
-    }
-
-    public LocalDate getFom() {
-        return fom;
-    }
-
-    public LocalDate getTom() {
-        return tom;
-    }
-
-    public List<YtelseBeregningsgrunnlagVedtak> getVedtak() {
-        return vedtak;
-    }
-
-    public abstract void mapSpesialverdier(YtelseGrunnlagBuilder builder);
-
-    public abstract boolean harArbeidsForhold();
-
-    public abstract List<YtelseBeregningsgrunnlagArbeidsforhold> getArbeidsforhold();
-
-    public abstract Arbeidskategori getArbeidskategori();
-
-    @Override
-    public String toString() {
-        return "YtelseBeregningsgrunnlag{" +
-            "type=" + type +
-            ", temaUnderkategori=" + temaUnderkategori +
-            ", identdato=" + identdato +
-            ", fom=" + fom +
-            ", tom=" + tom +
-            '}';
-    }
-}
