--- domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/ytelse/infotrygd/beregningsgrunnlag/YtelseBeregningsgrunnlagSykepenger.java
@@ -1,30 +1,0 @@
-package no.nav.foreldrepenger.abakus.registerdata.ytelse.infotrygd.beregningsgrunnlag;
-
-import java.math.BigDecimal;
-
-import no.nav.foreldrepenger.abakus.domene.iay.YtelseGrunnlagBuilder;
-import no.nav.foreldrepenger.abakus.kodeverk.YtelseType;
-import no.nav.foreldrepenger.abakus.vedtak.domene.TemaUnderkategori;
-import no.nav.tjeneste.virksomhet.infotrygdberegningsgrunnlag.v1.informasjon.Sykepenger;
-
-public class YtelseBeregningsgrunnlagSykepenger extends YtelseBeregningsgrunnlagPeriodeYtelse {
-
-    private BigDecimal inntektsgrunnlagProsent;
-
-    YtelseBeregningsgrunnlagSykepenger(Sykepenger sykepenger, TemaUnderkategori tuk) {
-        super(YtelseType.SYKEPENGER, tuk, sykepenger);
-        if (sykepenger.getInntektsgrunnlagProsent() != null) {
-            inntektsgrunnlagProsent = new BigDecimal(sykepenger.getInntektsgrunnlagProsent());
-        }
-    }
-
-    public BigDecimal getInntektsgrunnlagProsent() {
-        return inntektsgrunnlagProsent;
-    }
-
-
-    @Override
-    public void mapSpesialverdier(YtelseGrunnlagBuilder builder) {
-        builder.medInntektsgrunnlagProsent(inntektsgrunnlagProsent);
-    }
-}
