--- domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/ytelse/infotrygd/sak/InfotrygdKilde.java
@@ -1,6 +1,0 @@
-package no.nav.foreldrepenger.abakus.registerdata.ytelse.infotrygd.sak;
-
-public enum InfotrygdKilde {
-    SAK,
-    VEDTAK
-}
