--- domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/ytelse/infotrygd/sak/InfotrygdPersonIkkeFunnetException.java
@@ -1,10 +1,0 @@
-package no.nav.foreldrepenger.abakus.registerdata.ytelse.infotrygd.sak;
-
-import no.nav.vedtak.exception.IntegrasjonException;
-import no.nav.vedtak.feil.Feil;
-
-public class InfotrygdPersonIkkeFunnetException extends IntegrasjonException {
-    public InfotrygdPersonIkkeFunnetException(Feil feil) {
-        super(feil);
-    }
-}
