--- domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/ytelse/infotrygd/sak/InfotrygdUgyldigInputException.java
@@ -1,10 +1,0 @@
-package no.nav.foreldrepenger.abakus.registerdata.ytelse.infotrygd.sak;
-
-import no.nav.vedtak.exception.IntegrasjonException;
-import no.nav.vedtak.feil.Feil;
-
-public class InfotrygdUgyldigInputException extends IntegrasjonException {
-    public InfotrygdUgyldigInputException(Feil feil) {
-        super(feil);
-    }
-}
