+++ domenetjenester/iay/src/main/resources/META-INF/pu-default.inntektarbeidytelse.orm.xml
@@ -80,7 +80,6 @@
     <entity class="no.nav.foreldrepenger.abakus.domene.iay.kodeverk.SkatteOgAvgiftsregelType" />
     <entity class="no.nav.foreldrepenger.abakus.domene.iay.kodeverk.Arbeidskategori" />
     <entity class="no.nav.foreldrepenger.abakus.domene.iay.kodeverk.ArbeidType" />
-    <entity class="no.nav.foreldrepenger.abakus.domene.iay.kodeverk.FagsystemUnderkategori" />
     <entity class="no.nav.foreldrepenger.abakus.domene.iay.kodeverk.InntektsKilde" />
     <entity class="no.nav.foreldrepenger.abakus.domene.iay.kodeverk.InntektsmeldingInnsendingsårsak" />
     <entity class="no.nav.foreldrepenger.abakus.domene.iay.kodeverk.InntektspostType" />
