+++ migreringer/src/main/resources/db/migration/defaultDS/1.0/V1.1_59__fjern_infotrygdws_rester.sql
@@ -1,0 +1,5 @@
+alter table IAY_RELATERT_YTELSE
+    drop constraint FK_YTELSE_6,
+    drop column kl_fagsystem_underkategori ,
+    drop column fagsystem_underkategori;
+
