+++ web/app-vtp.properties
@@ -7,8 +7,6 @@ Kodeverk.v2.url=https://localhost:8063/soap/kodeverk/ws/Kodeverk/v2
 MeldekortUtbetalingsgrunnlag.v1.url=https://localhost:8063/soap/ail_ws/MeldekortUtbetalingsgrunnlag_v1
 SigrunRestBeregnetSkatt.url=https://localhost:8063/rest
 hentinntektlistebolk.url=https://localhost:8063/rest/inntektskomponenten-ws/rs/api/v1/hentinntektlistebolk
-InfotrygdSak.v1.url=https://localhost:8063/soap/infotrygd-ws/InfotrygdSak/v1
-InfotrygdBeregningsgrunnlag.v1.url=https://localhost:8063/soap/infotrygd-ws/InfotrygdBeregningsgrunnlag/v1
 fpabakus.it.fp.grunnlag.url=https://localhost:8063/rest/infotrygd/grunnlag/foreldrepenger
 fpabakus.it.svp.grunnlag.url=https://localhost:8063/rest/infotrygd/grunnlag/svangerskapspenger
 fpabakus.it.sp.grunnlag.url=https://localhost:8063/rest/infotrygd/grunnlag/sykepenger
