--- web/src/main/java/no/nav/foreldrepenger/abakus/app/selftest/checks/InfotrygdSakWebServiceHealthCheck.java
@@ -1,37 +1,0 @@
-package no.nav.foreldrepenger.abakus.app.selftest.checks;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
-import no.nav.vedtak.felles.integrasjon.infotrygdsak.InfotrygdSakSelftestConsumer;
-
-@ApplicationScoped
-public class InfotrygdSakWebServiceHealthCheck extends WebServiceHealthCheck {
-
-    private InfotrygdSakSelftestConsumer infotrygdSakSelftestConsumer;
-
-    InfotrygdSakWebServiceHealthCheck(){
-        //For CDI proxy
-    }
-
-    @Inject
-    public InfotrygdSakWebServiceHealthCheck(InfotrygdSakSelftestConsumer infotrygdSakSelftestConsumer){
-        this.infotrygdSakSelftestConsumer = infotrygdSakSelftestConsumer;
-    }
-
-    @Override
-    protected void performWebServiceSelftest() {
-        infotrygdSakSelftestConsumer.ping();
-    }
-
-    @Override
-    protected String getDescription() {
-        return "Test av web service InfotrygdSak";
-    }
-
-    @Override
-    protected String getEndpoint() {
-        return infotrygdSakSelftestConsumer.getEndpointUrl();
-    }
-
-}
